Move Expr-specific code in Basic.sort_key() to Expr
