+++ sympy/core/basic.py
@@ -382,51 +382,8 @@ class Basic(AssumeMeths):
         [x**(-2), 1/x, x**(1/4), x**(1/2), x, x**(3/2), x**2]
 
         """
-        from sympy.core import S
-        def rmap(args):
-            """Recursively map a tree. """
-            result = []
-
-            for arg in args:
-                if isinstance(arg, Basic):
-                    arg = arg.sort_key(order=order)
-                elif hasattr(arg, '__iter__'):
-                    arg = rmap(arg)
-
-                result.append(arg)
-
-            return tuple(result)
-
-        try:
-            coeff, expr = self.as_coeff_Mul()
-        except AttributeError:
-            return self.class_key(), (len(self.args), self.args), S.One.sort_key(), S.One
-        else:
-            if expr.is_Pow:
-                expr, exp = expr.args
-            else:
-                expr, exp = expr, S.One
-
-            if expr.is_Atom:
-                if expr.is_Symbol:
-                    args = (str(expr),)
-                else:
-                    args = (expr,)
-            else:
-                if expr.is_Add:
-                    args = expr.as_ordered_terms(order=order)
-                else:
-                    args = expr.args
-
-                args = rmap(args)
-
-                if expr.is_Mul:
-                    args = sorted(args)
-
-            args = (len(args), args)
-            exp = exp.sort_key(order=order)
-
-            return expr.class_key(), args, exp, coeff
+        from sympy.core.singleton import S
+        return self.class_key(), (len(self.args), self.args), S.One.sort_key(), S.One
 
 
     def __eq__(self, other):
